Remove short phrases from the game description
Update the game description to no longer state that short phrases are possible guesses. Further down, explicitly declare that the secret word will not be a pair of words, implicitly this applies to guesses too.

That explicit info is further down because I think this minor nitpick of what counts as a word should not be front and center - it pretty much goes without saying, and only worth stating as a disclaimer.